+++ core/UnitTests/Editor/SceneGraphControllerTest.cs
@@ -1,10 +1,11 @@
 ﻿using System.Collections.Generic;
 using NUnit.Framework;
 using UnityEngine;
-using WorldWizards.core.controller.level;
-using WorldWizards.core.controller.level.utils;
+using WorldWizards.core.controller.resources;
 using WorldWizards.core.entity.coordinate;
 using WorldWizards.core.entity.gameObject;
+using WorldWizards.core.entity.gameObject.utils;
+using WorldWizards.core.experimental;
 using WorldWizards.core.file.utils;
 using WorldWizards.core.manager;
