--- core/controller/level.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 56e0944a2e2fc4a9eb47f8bcefad1557
-folderAsset: yes
-timeCreated: 1503679003
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
