--- core/controller/level/utils.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: ccbcb8b08a838401e9800dcec6031217
-folderAsset: yes
-timeCreated: 1503679003
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
