+++ core/controller/resources/WWAssetBundleController.cs.meta
