+++ core/entity/coordinate/IntVector3.cs.meta
