+++ core/viveControllers/ControllerMakeCube.cs
@@ -1,8 +1,8 @@
 ﻿using UnityEngine;
-using WorldWizards.core.controller.level.utils;
 using WorldWizards.core.entity.coordinate;
 using WorldWizards.core.entity.coordinate.utils;
 using WorldWizards.core.entity.gameObject;
+using WorldWizards.core.entity.gameObject.utils;
 using WorldWizards.core.manager;
 using WorldWizards.SteamVR.Scripts;
